backend/requirements.txt
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
flake8==4.0.1
pytest-django==4.4.0

